Update EndpointOut to allow arbitrary size reads
This updates the EndpointOut::read() API to support reads larger than
the maximum packet size.  A new EndpointOutSinglePacket helper trait
was added to make it easy for driver implementations to provide this
functionality even if they implement reading one packet at a time.
This adds new unit tests to verify the behavior of
EndpointOutSinglePacket.

One caveat about the API here is the fact that dropping a read() call
may lose data.  At the moment the existing driver implementations
currently do not lose data on a dropped read, but this adds extra read
overhead.  e.g., the embassy-stm32 implementation always reads data into
an intermediate buffer rather than being able to read directly into the
caller's buffer.  This extra buffer allows it to avoid losing data if
the read operation is dropped.  If read calls can only read up to 1
packet at a time, then it is more feasible to have enough buffer space
to do this, if you are willing to pay the overhead of the extra data.
However, with arbitrary sized reads it becomes trickier to ensure that
you always have enough buffer space to save an arbitrary amount of data
between a dropped read() operation and the next call to read().

Because read operations might lose data when they are dropped, it seems
like callers would potentially need to reset the bus to recover after a
dropped read (except perhaps for isochronous endpoints).

If needed, it wouldn't be too hard for drivers to implement an API that
allows canceling an in-progress read operation, adding a function call
that causes any in-progress read on a particular endpoint to return
early, and to include the partial read length when it returns.  However,
there is an API design question for where to put this API, since we
couldn't put it on the Endpoint object if the Endpoint is already
borrowed for the duration of the read operation.